+++ app/service/store/internal/model/store_address.go
@@ -4,9 +4,9 @@ type CommonStoreAddress struct {
 	ProvinceName string `json:"province_name"`
 	CityName string `json:"city_name"`
 	AreaName string `json:"area_name"`
-	ProvinceCode int64 `json:"province_code"`
-	CityCode int64 `json:"city_code"`
-	AreaCode int64  `json:"area_code"`
+	ProvinceCode int `json:"province_code"`
+	CityCode int `json:"city_code"`
+	AreaCode int  `json:"area_code"`
 	Address string  `json:"address"` // 详细地址
 	Gcj02 string    `json:"gcj_02"` // 高德坐标
 	GaodeId string  `json:"gaode_id"` // 高德id
